v5.2.1 "Knowledge System with Full Service & Error Visibility"
Features:
- [SERVICES] section: Total/Active/Inactive/Enabled/Disabled/Masked/Failed
- [ERRORS] section: Failed services and recent log errors (24h)
- [LOG SCAN] section: Daemon log scanner state and statistics
- Extended [INVENTORY] with packages, commands, services, configs, logs
- [DISCOVERY] section for priority scan status in focused views
- [SERVICE RELATIONS] for packages/commands showing related systemd units
- LogScanState tracking for daemon log scanner
- ServiceIndex extended counts (inactive, enabled, disabled)
- recent_errors_24h() and errors_since() for ErrorIndex
- find_services_using_executable() for service relations
- 8 new v5.2.1 invariant tests
- Clippy warnings fixed across workspace

+++ crates/annad/tests/lifecycle_correctness_tests.rs
@@ -277,7 +277,7 @@ fn test_post_reset_lifecycle_clean() {
     // Verify new lifecycle is clean and separate
     let phase3_state = ExperienceSnapshot::capture(&paths);
     assert_eq!(
-        phase3_state.telemetry_line_count as u64,
+        phase3_state.telemetry_line_count,
         new_count,
         "Phase 3 should have exactly {} telemetry entries (the new questions answered)",
         new_count
